refactor: Added helper methods <contructEtherAmount, toAmountInDouble> in utils/extensions of mirai_web3

+++ packages/mirai_web3/lib/models/transaction_details.dart
@@ -13,7 +13,7 @@ class TransactionDetails {
 
   final String senderAddress;
   final String receiverAddress;
-  final BigInt amount;
+  final double amount;
   final bool received;
   final String tranHash;
   final String tokenAddress;
